Unicode/UTF8 related fixes in CTextBox/Console (yet incomplete)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@549 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/CssParser.h
@@ -6,6 +6,8 @@
 #ifndef __CSS_PARSER_H__
 #define __CSS_PARSER_H__
 
+// TODO: recode everything here! (use std::string)
+
 typedef struct property_s {
 	char		*sName;
 	char		*sValue;
@@ -32,8 +34,8 @@ private:
 	// Attributes
 	node_t	*tNodes;
 
-	int		iPos;
-	int		iLength;
+	size_t	iPos;
+	size_t	iLength;
 	char	*sData;
 
 	// Internal methods
